// tutor/src/payload-types.ts
export interface Course {
  id: string;
  courseName: string;
  tuition?: number | null;
  total_hours?: number | null;
  skills?: ('reading' | 'writing' | 'speaking' | 'listening' | 'grammar' | 'vocabulary')[] | null;
  type?: ('online' | 'offline') | null;
  thumbnail?: (string | null) | Media;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: string;
  alt?: string | null;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
// ...
export interface Admin {
  id: string;
  department: 'ceo' | 'hoc_vu' | 'hoc_thuat' | 'sale' | 'marketing' | 'hr' | 'ke_toan' | '25';
  role?:
    | (
        | 'admin'
        | 'hoc_vu_manager'
        | 'hoc_vu_executive'
        | 'academic_manager'
        | 'academic_executive'
        | 'teacher_part_time'
        | 'teacher_full_time'
        | 'wow_manager'
        | 'wow_executive'
        | 'sale_manager'
        | 'sale_executive'
        | 'marketing_manager'
        | 'marketing_executive'
        | 'designer'
        | 'seo'
        | 'digital_marketing'
        | 'hr_manager'
        | 'hr_generalist'
        | 'ke_toan'
      )
    | null;
  fullName: string;
  phone: string;
  birthday?: string | null;
  working_date?: string | null;
  address?: string | null;
  avatar?: (string | null) | Media;
  /**
   * Dùng để gửi cho học viên khi học WOW trực tuyến
   */
  meetingLink?: string | null;
  availability?:
    | {
        /**
         * Chọn ngày rảnh của WOW để học sinh có thể book. Ví dụ bắt đầu vào ngày 01/01/2025 9:00 và kết thúc vào ngày 01/01/2025 17:00. Có thể chọn nhiều khung giờ khác nhau cùng ngày nếu như ngày đó WOW rảnh vào nhiều khung giờ
         */
        start_date: string;
        end_date: string;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
  email?: string | null;
  username: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'admins';
}
// ...
export interface Lead {
  id: string;
  displayTitle?: string | null;
  title?: string | null;
  fullName?: string | null;
  phone?: string | null;
  email?: string | null;
  gender?: ('male' | 'female') | null;
  'parent name'?: string | null;
  'parent phone'?: string | null;
  'parent email'?: string | null;
  date_of_birth?: string | null;
  link_facebook?: string | null;
  address?: string | null;
  type?: ('student' | 'college_student' | 'office_worker' | 'parent' | 'other') | null;
  target?: number | null;
  wow_quota?: number | null;
  purpose?: string | null;
  branch?: (string | null) | Branch;
  source:
    | 'facebook'
    | 'google'
    | 'hotline'
    | 'walking'
    | 'referral'
    | 'seeding'
    | 'school_tour'
    | 'sponsorship'
    | 're_enroll'
    | 'from_website'
    | 'other';
  saleInCharge?: (string | null) | Admin;
  status?:
    | (
        | 'waiting_for_call'
        | 'test_booked'
        | 'test_not_booked'
        | 'potential'
        | 'not_potential'
        | 'paid_case'
        | 'failed_case'
        | 'consider'
      )
    | null;
  courses?: string | null;
  coupon?: (string | Coupon)[] | null;
  notes?:
    | {
        channel_interaction?: ('zalo' | 'phone' | 'fanpage' | 'hotline' | 'direct_message' | 'online') | null;
        appointment_date?: string | null;
        status?:
          | (
              | 'waiting_for_call'
              | 'test_booked'
              | 'test_not_booked'
              | 'potential'
              | 'not_potential'
              | 'paid_case'
              | 'failed_case'
              | 'consider'
            )
          | null;
        message?: string | null;
        createdAt?: string | null;
        createdBy?: (string | null) | Admin;
        id?: string | null;
      }[]
    | null;
  chats?:
    | {
        createdBy?: (string | null) | Admin;
        message?: string | null;
        createdAt?: string | null;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "branches".
 */
export interface Branch {
  id: string;
  name: string;
  address: string;
  phone: string;
  rooms: (string | Room)[];
  status?: ('active' | 'unactive') | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "rooms".
 */
export interface Room {
  id: string;
  name: string;
  capacity?: number | null;
  status?: ('active' | 'unactive') | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "coupons".
 */
export interface Coupon {
  id: string;
  name: string;
  discount: number;
  type: 'money' | 'percent';
  startDate: string;
  endDate: string;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "band_score".
 */
export interface BandScore {
  id: string;
  score: string;
  skillLevel?: string | null;
  description?: string | null;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "pages".
 */
export interface Page {
  id: string;
  title: string;
  slug?: string | null;
  slugLock?: boolean | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "users".
 */
export interface User {
  id: string;
  avatar?: (string | null) | Media;
  role: 'user';
  lead?: (string | null) | Lead;
  violationStatus?: ('none' | 'warning_1' | 'warning_2' | 'terminated') | null;
  displayTitle?: string | null;
  updatedAt: string;
  createdAt: string;
  email?: string | null;
  username: string;
  resetPasswordToken?: string | null;
  resetPasswordExpiration?: string | null;
  salt?: string | null;
  hash?: string | null;
  loginAttempts?: number | null;
  lockUntil?: string | null;
  sessions?:
    | {
        id: string;
        createdAt?: string | null;
        expiresAt: string;
      }[]
    | null;
  password?: string | null;
  collection: 'users';
}
// ...
export interface Class {
  id: string;
  name: string;
  type_class?: ('group' | 'one_to_one') | null;
  course?: (string | null) | Course;
  startDate: string;
  /**
   * Học vào những ngày nào trong tuần. Ví dụ: Thứ 2, Thứ 4, Thứ 6. Được chọn nhiều
   */
  days_of_week?: ('1' | '2' | '3' | '4' | '5' | '6' | '0')[] | null;
  session_time?: string | null;
  total_hours?: number | null;
  time_per_session?: number | null;
  students?: (string | User)[] | null;
  status_class?: ('opening' | 'active' | 'closed') | null;
  link_zoom?: string | null;
  branch?: (string | null) | Branch;
  room?: (string | null) | Room;
  teachers?:
    | {
        /**
         * Chọn giáo viên phụ trách cho lớp. Được chọn nhiều. Chỉ hiển thị giáo viên full-time và part-time
         */
        teacher?: (string | null) | Admin;
        skill?: ('reading' | 'writing' | 'speaking' | 'listening' | 'grammar' | 'vocabulary')[] | null;
        /**
         * Dạy vào những ngày nào trong tuần. Ví dụ: Thứ 2, Thứ 4, Thứ 6. Được chọn nhiều
         */
        days_of_week?: ('1' | '2' | '3' | '4' | '5' | '6' | '0')[] | null;
        session_time?: string | null;
        duration?: number | null;
        id?: string | null;
      }[]
    | null;
  sessions?:
    | {
        date: string;
        teacher?: (string | null) | Admin;
        session_time?: string | null;
        duration?: number | null;
        id?: string | null;
      }[]
    | null;
  roadmaps?:
    | {
        skill?: ('reading' | 'writing' | 'speaking' | 'listening' | 'grammar' | 'vocabulary') | null;
        content: string;
        id?: string | null;
      }[]
    | null;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface PeriodicTest {
  id: string;
  title: string;
  description?: string | null;
  type: 'homework' | 'extra_homework' | 'mini_test' | 'mid_term' | 'final_term' | 'bank' | 'entrance_test';
  tests: (string | Test)[];
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
// ...
export interface PeriodicTestAttempt {
  id: string;
  user: string | User;
  test: string | PeriodicTest;
  type: 'homework' | 'extra_homework' | 'mini_test' | 'mid_term' | 'final_term' | 'bank' | 'entrance_test';
  status: 'pending' | 'completed' | 'in_progress';
  class?: (string | null) | Class;
  session?: number | null;
  score?: (string | null) | BandScore;
  startedAt?: string | null;
  completedAt?: string | null;
  duration?: string | null;
  mode?: ('practice' | 'simulation') | null;
  part?: string | null;
  time?: number | null;
  statusDetails?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  answerSheet?:
    | {
        [k: string]: unknown;
      }
    | unknown[]
    | string
    | number
    | boolean
    | null;
  updatedAt: string;
  createdAt: string;
}